Refactor backend folders and files structure.
