+++ backend/constants/deviceTypes.js
@@ -1,0 +1,4 @@
+export const DEVICE_TYPES = {
+  DEVICE: 'device',
+  GATEWAY: 'gateway',
+};
