+++ backend/middleware/auth.js
@@ -1,0 +1,29 @@
+export const verifyToken = async (req, res, next) => {
+  const authHeader = req.headers.authorization;
+
+  if (!authHeader || !authHeader.startsWith('Bearer ')) {
+    return res.status(401).json({ error: 'No token provided' });
+  }
+
+  const token = authHeader.split(' ')[1];
+
+  try {
+    const userInfoResponse = await fetch(
+      'https://www.googleapis.com/oauth2/v3/userinfo',
+      {
+        headers: { Authorization: `Bearer ${token}` },
+      },
+    );
+
+    if (!userInfoResponse.ok) {
+      return res.status(401).json({ error: 'Invalid token' });
+    }
+
+    const userInfo = await userInfoResponse.json();
+    req.user = userInfo;
+    next();
+  } catch (error) {
+    console.error('Token verification error:', error);
+    res.status(401).json({ error: 'Invalid or expired token' });
+  }
+};
